Rename CoursePage to ViewerPage
