+++ app/components/ApplicationPage/TopicPage/ViewerPage/index.test.js
@@ -1,0 +1,37 @@
+// @flow
+
+import _ from 'lodash';
+import * as React from 'react';
+import { shallow } from 'enzyme';
+
+import { dummyProviderProps } from 'lib/testResources';
+
+import { PureViewerPage } from '.';
+
+describe(`ViewerPage`, (): void => {
+
+  let dummyTopicId: string;
+
+  beforeEach((): void => {
+    dummyTopicId = 'dummyTopicId';
+  });
+
+  it(`renders without errors`, (): void => {
+    const fixedRouterProps = _.set(_.cloneDeep(dummyProviderProps.routerProps), 'match.params.topicId', dummyTopicId);
+
+    const enzymeWrapper = shallow(
+      <PureViewerPage {...fixedRouterProps} />,
+    );
+
+    expect(enzymeWrapper.isEmptyRender()).toEqual(false);
+  });
+
+  it(`renders NULL when match.params.id is NULL`, (): void => {
+    const enzymeWrapper = shallow(
+      <PureViewerPage {...dummyProviderProps.routerProps} />,
+    );
+
+    expect(enzymeWrapper.isEmptyRender()).toEqual(true);
+  });
+
+});
